feat: use `--mount`s instead of COPY
Rather than creating a layer, use a readonly (by default) filesystem
mount to represent the local source tree or "install" various tools.

The "why" on the former is a little unclear as yet; for the latter, it's
cheaper to "borrow" the tool that way rather than needing to create a
layer for it.