fix: llb.IncludePatterns don't support `**`
Prior to this change, we were including every file from the project
root (including non-go file), but weren't including any go files
from deeper directories than the root.

llb.IncludePatterns ultimately use `filepath.Match` to check whether a
path should be included, and it doesn't support `**` as an operator. So
the intent was to say "match any number of separated segments, including
zero, as long as the last component matches `*.go`, but that's not what
was happening.